broke apart connection
broke apart the connection class into multiple partial classes each of which represents a specific communication style supported by KubeMQ.
