+++ Contract/ConnectionOptions.cs
@@ -1,5 +1,6 @@
 ﻿using Grpc.Core;
 using KubeMQ.Contract.Interfaces;
+using KubeMQ.Contract.SDK;
 using KubeMQ.Contract.SDK.Grpc;
 using Microsoft.Extensions.Logging;
 
@@ -86,6 +87,45 @@ namespace KubeMQ.Contract
         public IConnection EstablishConnection(IGlobalMessageEncoder? globalMessageEncoder=null,IGlobalMessageEncryptor? globalMessageEncryptor=null)
         {
             return new Connection(this,globalMessageEncoder,globalMessageEncryptor);
+        }
+
+        /// <summary>
+        /// Called to use the Current Options to establish a Pub/Sub connection to the KubeMQ server.
+        /// </summary>
+        /// <param name="globalMessageEncoder">If desired, an encoder can be specified here and will be used to encode message bodies as the default.  
+        /// A type specific encoder can be specified to override this for that particular type of message.</param>
+        /// <param name="globalMessageEncryptor">If desired, an encryptor can be specified here and will be used to secure the message bodies as the default.  
+        /// A type specific encryptor can be specified to override this for that particular type of message.</param>
+        /// <returns></returns>
+        public IPubSubConnection EstablishPubSubConnection(IGlobalMessageEncoder? globalMessageEncoder = null, IGlobalMessageEncryptor? globalMessageEncryptor = null)
+        {
+            return new Connection(this, globalMessageEncoder, globalMessageEncryptor);
+        }
+
+        /// <summary>
+        /// Called to use the Current Options to establish a RPC connection to the KubeMQ server.
+        /// </summary>
+        /// <param name="globalMessageEncoder">If desired, an encoder can be specified here and will be used to encode message bodies as the default.  
+        /// A type specific encoder can be specified to override this for that particular type of message.</param>
+        /// <param name="globalMessageEncryptor">If desired, an encryptor can be specified here and will be used to secure the message bodies as the default.  
+        /// A type specific encryptor can be specified to override this for that particular type of message.</param>
+        /// <returns></returns>
+        public IRPCConnection EstablishRPCConnection(IGlobalMessageEncoder? globalMessageEncoder = null, IGlobalMessageEncryptor? globalMessageEncryptor = null)
+        {
+            return new Connection(this, globalMessageEncoder, globalMessageEncryptor);
+        }
+
+        /// <summary>
+        /// Called to use the Current Options to establish a Queue connection to the KubeMQ server.
+        /// </summary>
+        /// <param name="globalMessageEncoder">If desired, an encoder can be specified here and will be used to encode message bodies as the default.  
+        /// A type specific encoder can be specified to override this for that particular type of message.</param>
+        /// <param name="globalMessageEncryptor">If desired, an encryptor can be specified here and will be used to secure the message bodies as the default.  
+        /// A type specific encryptor can be specified to override this for that particular type of message.</param>
+        /// <returns></returns>
+        public IQueueConnection EstablishQueueConnection(IGlobalMessageEncoder? globalMessageEncoder = null, IGlobalMessageEncryptor? globalMessageEncryptor = null)
+        {
+            return new Connection(this, globalMessageEncoder, globalMessageEncryptor);
         }
     }
 }
